// src/resources/data/projects/table/columns.ts
export class Columns extends APIResource {
  /**
   * Add Column
   *
   * @example
   * ```ts
   * const column =
   *   await client.data.projects.table.columns.create(
   *     'tableName',
   *     { projectId: 'projectId' },
   *   );
   *
   * const content = await column.blob();
   * console.log(content);
   * ```
   */
  create(tableName: string, params: ColumnCreateParams, options?: RequestOptions): APIPromise<Response> {
    const { projectId, ...body } = params;
    return this._client.post(path`/data/projects/${projectId}/table/${tableName}/columns`, {
      body,
      ...options,
      headers: buildHeaders([{ Accept: '*/*' }, options?.headers]),
      __binaryResponse: true,
    });
  }

  /**
   * Update Column
   *
   * @example
   * ```ts
   * const column =
   *   await client.data.projects.table.columns.update(
   *     'columnName',
   *     { projectId: 'projectId', tableName: 'tableName' },
   *   );
   *
   * const content = await column.blob();
   * console.log(content);
   * ```
   */
  update(columnName: string, params: ColumnUpdateParams, options?: RequestOptions): APIPromise<Response> {
    const { projectId, tableName, ...body } = params;
    return this._client.put(path`/data/projects/${projectId}/table/${tableName}/columns/${columnName}`, {
      body,
      ...options,
      headers: buildHeaders([{ Accept: '*/*' }, options?.headers]),
      __binaryResponse: true,
    });
  }

  /**
   * See Data -- Catalog / List Columns for more details.
   *
   * @example
   * ```ts
   * const columns =
   *   await client.data.projects.table.columns.list(
   *     'tableName',
   *     { projectId: 'projectId' },
   *   );
   *
   * const content = await columns.blob();
   * console.log(content);
   * ```
   */
  list(tableName: string, params: ColumnListParams, options?: RequestOptions): APIPromise<Response> {
    const { projectId } = params;
    return this._client.get(path`/data/projects/${projectId}/table/${tableName}/columns`, {
      ...options,
      headers: buildHeaders([{ Accept: '*/*' }, options?.headers]),
      __binaryResponse: true,
    });
  }

  /**
   * Delete Column
   *
   * @example
   * ```ts
   * await client.data.projects.table.columns.delete(
   *   'columnName',
   *   { projectId: 'projectId', tableName: 'tableName' },
   * );
   * ```
   */
  delete(columnName: string, params: ColumnDeleteParams, options?: RequestOptions): APIPromise<void> {
    const { projectId, tableName } = params;
    return this._client.delete(path`/data/projects/${projectId}/table/${tableName}/columns/${columnName}`, {
      ...options,
      headers: buildHeaders([{ Accept: '*/*' }, options?.headers]),
    });
  }
}

export interface ColumnCreateParams {
  /**
   * Path param:
   */
  projectId: string;

  /**
   * Body param: The data type of the column.
   */
  dataType?: string;

  /**
   * Body param: The default value of the column.
   */
  defaultValue?: string;

  /**
   * Body param: The display name of the column.
   */
  displayName?: string;

  /**
   * Body param: The not null flag for the column.
   */
  isNotNull?: boolean;

  /**
   * Body param: The unique flag for the column.
   */
  isUnique?: boolean;

  /**
   * Body param: The name of the column.
   */
  name?: string;
}

export interface ColumnUpdateParams {
  /**
   * Path param:
   */
  projectId: string;

  /**
   * Path param:
   */
  tableName: string;

  /**
   * Body param: The data type of the column.
   */
  dataType?: string;

  /**
   * Body param: The default value of the column.
   */
  defaultValue?: string;

  /**
   * Body param: The display name of the column.
   */
  displayName?: string;

  /**
   * Body param: The not null flag for the column.
   */
  isNotNull?: boolean;

  /**
   * Body param: The unique flag for the column.
   */
  isUnique?: boolean;

  /**
   * Body param: The name of the column.
   */
  name?: string;
}

export interface ColumnListParams {
  projectId: string;
}

export interface ColumnDeleteParams {
  projectId: string;

  tableName: string;
}

export declare namespace Columns {
  export {
    type ColumnCreateParams as ColumnCreateParams,
    type ColumnUpdateParams as ColumnUpdateParams,
    type ColumnListParams as ColumnListParams,
    type ColumnDeleteParams as ColumnDeleteParams,
  };
}